[alpha_factory] add archive page
