+++ src/interface/web_client/package.json
@@ -12,12 +12,15 @@
   "dependencies": {
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "plotly.js-dist": "^2.24.2"
+    "plotly.js-dist": "^2.24.2",
+    "react-router-dom": "^6.17.0",
+    "vue": "^3.3.4"
   },
   "devDependencies": {
     "@types/react": "^18.2.0",
     "@types/react-dom": "^18.2.0",
     "@vitejs/plugin-react": "^4.2.0",
+    "@vitejs/plugin-vue": "^5.0.0",
     "typescript": "^5.3.0",
     "vite": "^4.2.0"
   }
