+++ src/interface/web_client/vite.config.ts
@@ -1,9 +1,10 @@
 // SPDX-License-Identifier: Apache-2.0
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
+import vue from '@vitejs/plugin-vue';
 
 export default defineConfig({
-  plugins: [react()],
+  plugins: [react(), vue()],
   root: '.',
   envPrefix: 'VITE_',
   build: {
